requirements.txt
appdirs==1.4.4
attrs==23.2.0
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.2.2
charset-normalizer==3.3.2
contourpy==1.2.0
cssselect==1.2.0
cycler==0.12.1
Cython==0.29.37
exceptiongroup==1.2.0
fake-useragent==1.4.0
fonttools==4.49.0
h11==0.14.0
hdbscan==0.8.33
idna==3.6
importlib-metadata==7.0.1
joblib==1.3.2
kiwisolver==1.4.5
llvmlite==0.42.0
lxml==5.1.0
matplotlib==3.8.3
numba==0.59.0
numpy==1.26.4
outcome==1.3.0.post0
packaging==23.2
pandas==2.2.1
parse==1.20.1
pillow==10.2.0
plotly==5.19.0
pyee==11.1.0
pynndescent==0.5.11
pyparsing==3.1.1
pyppeteer==2.0.0
pyquery==2.0.0
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2024.1
requests==2.31.0
requests-html==0.10.0
scikit-learn==1.4.1.post1
scipy==1.12.0
selenium==4.18.1
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
tenacity==8.2.3
threadpoolctl==3.3.0
tqdm==4.66.2
trio==0.24.0
trio-websocket==0.11.1
typing_extensions==4.10.0
tzdata==2024.1
umap-learn==0.5.5
urllib3==1.26.18
w3lib==2.1.2
websockets==10.4
wsproto==1.2.0
zipp==3.17.0
